+++ src/main.tsx
@@ -2,7 +2,7 @@ import './index.css';
 
 import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
-import { router } from './app/router/router';
+import { router } from './router/router';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <RouterProvider router={router} />,
