+++ src/router/router.tsx
@@ -4,7 +4,7 @@ import {
   Route,
 } from 'react-router-dom';
 import App from '../App';
-import ResultDetails from '../../components/Results/ResultDetails/ResultDetails';
+import ResultDetails from '../components/Results/ResultDetails/ResultDetails';
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
